feat(@clayui/language-picker): LPD-45714 Move component to clay-core
