+++ packages/clay-core/package.json
@@ -27,6 +27,7 @@
 		"@clayui/button": "^3.116.0",
 		"@clayui/form": "^3.125.0",
 		"@clayui/icon": "^3.111.0",
+		"@clayui/label": "^3.111.0",
 		"@clayui/layout": "^3.111.0",
 		"@clayui/link": "^3.111.0",
 		"@clayui/loading-indicator": "^3.111.0",
