--- packages/clay-language-picker/BREAKING.md
@@ -1,1 +1,0 @@
-# List of Breaking Changes for 4.x
