--- packages/clay-language-picker/CHANGELOG.md
@@ -1,4 +1,0 @@
-# Change Log
-
-All notable changes to this project will be documented in this file.
-See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
