--- packages/clay-language-picker/README.md
@@ -1,19 +1,0 @@
-# `@clayui/language-picker`
-
-A Picker used to select different languages. It also has the possibility of being used to select translation languages.
-
--   [Documentation](https://clayui.com/docs/components/language-picker.html)
--   [Changelog](./CHANGELOG.md)
--   [Breaking change schedule](./BREAKING.md)
-
-## Install
-
-Run `yarn`
-
-```shell
-yarn add @clayui/language-picker
-```
-
-## Contribute
-
-We'd love to get contributions from you! Please, check our [Contributing Guidelines](https://github.com/liferay/clay/blob/master/CONTRIBUTING.md) to see how you can help us improve.
