--- packages/clay-language-picker/package.json
@@ -1,42 +1,0 @@
-{
-	"name": "@clayui/language-picker",
-	"version": "1.0.0",
-	"description": "ClayLanguagePicker component",
-	"license": "BSD-3-Clause",
-	"repository": "https://github.com/liferay/clay",
-	"engines": {
-		"node": ">=0.12.0",
-		"npm": ">=3.0.0"
-	},
-	"main": "lib/index.js",
-	"types": "lib/index.d.ts",
-	"ts:main": "src/index.tsx",
-	"files": [
-		"lib"
-	],
-	"scripts": {
-		"build": "cross-env NODE_ENV=production babel src --root-mode upward --out-dir lib --extensions .ts,.tsx",
-		"buildTypes": "cross-env NODE_ENV=production tsc --project ./tsconfig.declarations.json",
-		"test": "jest --config ../../jest.config.js"
-	},
-	"keywords": [
-		"clay",
-		"react"
-	],
-	"dependencies": {
-		"@clayui/core": "^3.124.0",
-		"@clayui/icon": "^3.111.0",
-		"@clayui/label": "^3.111.0",
-		"@clayui/layout": "^3.111.0",
-		"@clayui/shared": "^3.119.0",
-		"classnames": "^2.2.6"
-	},
-	"peerDependencies": {
-		"@clayui/css": "3.x",
-		"react": "^16.12.0",
-		"react-dom": "^16.12.0"
-	},
-	"browserslist": [
-		"extends browserslist-config-clay"
-	]
-}
