--- packages/clay-language-picker/tsconfig.declarations.json
@@ -1,8 +1,0 @@
-{
-	"compilerOptions": {
-		"baseUrl": ".",
-		"declarationDir": "./lib"
-	},
-	"extends": "../../tsconfig.declarations.json",
-	"include": ["src"]
-}
